Update javascript-leetcode-valid-anagram.md
Hi!

I have found a bug in your algorithm. 

For example with the following input data:
s = "ab"
t = "ba"

We'll get false but we should get true.

This is because in the second loop we take the element by the index of the string s, while we should take the element by the index of the string t. 
Otherwise, we just subtract the extra number of characters that are in the string and stored in the hash table, since the arrangement of elements in strings is different.
Then the algorithm will work like you described.